pyproject.toml
[project]
name = "pr-reviewer-mcp"
version = "0.1.0"
description = "GitHub PR Review MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.12.3",
    "notion-client>=2.4.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
